+++ test-files/test_lua.lua
@@ -1,14 +1,5 @@
--- Single line comment in Lua
 
---[[
-Multi-line
-Lua comment
-]]
 
-print("Hello, Lua!") -- End of line comment
+print("Hello, Lua!")
 
---[[
-    Block comment with indentation
-    that should be removed
-]]
-local x = 42 -- Variable with comment
+local x = 42