+++ test-files/test_ruby.rb
@@ -1,14 +1,7 @@
-# This is a single-line comment in Ruby
 
-=begin
-This is a
-multi-line Ruby comment
-that should be removed
-=end
 
-puts "Hello, Ruby!" # This prints a message
+puts "Hello, Ruby!"
 
-x = 42 # Variable assignment with comment
+x = 42
 
 #! /usr/bin/env ruby
-# The above is a shebang and should be preserved